extract keypads
